Deduplicate and move configuration vars out of experiment.mk
Move the variables that are important for the Terraform configuration
back to the Makefile. Deduplicate variable assignments in experiment.mk
that are also present in Makefile. You gotta make up your mind on
the single place where these need to be defined, and I don't think
the experiment parameters are it.