README: split steps between setup and test run
The QA method document in the cometbft repository currently refers to
numbered ranges of steps in this README as setup and run sequences
that the tester needs to perform at different stages.
When the list of steps changes as it did in #35, updating these
hardcoded references is tedious and error-prone (case in point,
I failed to update quite a few places).

The solution is to do to these human-consumed texts what programmers do
to machine code and introduce subroutines!
The step sequences are split under headers that can be "invoked" by name
and anchored in the HTML rendering.